+++ timeutil/timeutil.go
@@ -23,6 +23,20 @@ import (
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 const (
+	NS     = time.Nanosecond
+	US     = time.Microsecond
+	MS     = time.Millisecond
+	SECOND = time.Second
+	MINUTE = time.Minute
+	HOUR   = time.Hour
+	DAY    = 24 * HOUR
+	WEEK   = 7 * DAY
+	YEAR   = 365 * DAY
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+const (
 	_SECOND int64 = 1
 	_MINUTE int64 = 60
 	_HOUR   int64 = 3600
@@ -66,8 +80,8 @@ func PrettyDurationInDays(d any) string {
 		return ""
 	}
 
-	if dur < 24*time.Hour {
-		dur = 24 * time.Hour
+	if dur < 24*HOUR {
+		dur = 24 * HOUR
 	}
 
 	days := int(dur.Hours()) / 24
@@ -255,18 +269,44 @@ func ParseDuration(dur string, defMod ...rune) (time.Duration, error) {
 
 // StartOfHour returns start of the hour
 func StartOfHour(t time.Time) time.Time {
+	if t.IsZero() {
+		return t
+	}
+
 	return time.Date(
 		t.Year(), t.Month(), t.Day(),
 		t.Hour(), 0, 0, 0, t.Location(),
 	)
 }
 
+// EndOfHour returns end of the hour
+func EndOfHour(t time.Time) time.Time {
+	if t.IsZero() {
+		return t
+	}
+
+	return StartOfHour(t).Add(HOUR).Add(-1 * time.Nanosecond)
+}
+
 // StartOfDay returns start of the day
 func StartOfDay(t time.Time) time.Time {
+	if t.IsZero() {
+		return t
+	}
+
 	return time.Date(
 		t.Year(), t.Month(), t.Day(),
 		0, 0, 0, 0, t.Location(),
 	)
+}
+
+// EndOfDay returns end of the day
+func EndOfDay(t time.Time) time.Time {
+	if t.IsZero() {
+		return t
+	}
+
+	return StartOfDay(t).Add(24 * HOUR).Add(-1 * time.Nanosecond)
 }
 
 // StartOfWeek returns the first day of the week
@@ -287,20 +327,55 @@ func StartOfWeek(t time.Time, firstDay time.Weekday) time.Time {
 	}
 }
 
+// EndOfWeek returns the last day of the week
+func EndOfWeek(t time.Time, firstDay time.Weekday) time.Time {
+	if t.IsZero() {
+		return t
+	}
+
+	return StartOfWeek(t, firstDay).AddDate(0, 0, 7).Add(-1 * time.Nanosecond)
+}
+
 // StartOfMonth returns the first day of the month
 func StartOfMonth(t time.Time) time.Time {
+	if t.IsZero() {
+		return t
+	}
+
 	return time.Date(
 		t.Year(), t.Month(), 1,
 		0, 0, 0, 0, t.Location(),
 	)
 }
 
+// EndOfMonth returns the last day of the month
+func EndOfMonth(t time.Time) time.Time {
+	if t.IsZero() {
+		return t
+	}
+
+	return StartOfMonth(t).AddDate(0, 1, 0).Add(-1 * time.Nanosecond)
+}
+
 // StartOfYear returns the first day of the year
 func StartOfYear(t time.Time) time.Time {
+	if t.IsZero() {
+		return t
+	}
+
 	return time.Date(
 		t.Year(), time.January, 1,
 		0, 0, 0, 0, t.Location(),
 	)
+}
+
+// EndOfYear returns the last day of the year
+func EndOfYear(t time.Time) time.Time {
+	if t.IsZero() {
+		return t
+	}
+
+	return StartOfYear(t).AddDate(1, 0, 0).Add(-1 * time.Nanosecond)
 }
 
 // PrevDay returns previous day date
@@ -382,6 +457,17 @@ func NextWeekend(t time.Time) time.Time {
 			return t
 		}
 	}
+}
+
+// IsWeekend returns true if given day is weekend (saturday or sunday)
+func IsWeekend(t time.Time) bool {
+	return t.Weekday() == time.Saturday || t.Weekday() == time.Sunday
+}
+
+// Until returns time until given moment
+func Until(t time.Time, mod time.Duration) int {
+	now := time.Now().In(t.Location())
+	return int(t.Sub(now) / mod)
 }
 
 // FromISOWeek returns date for given week number in given year
@@ -761,40 +847,40 @@ func getPrettySimpleDuration(dur time.Duration) string {
 
 func getPrettyShortDuration(d time.Duration, separator string) string {
 	switch {
-	case d >= 24*time.Hour:
+	case d >= DAY:
 		return fmt.Sprintf(
 			"%.0f"+separator+"d",
-			formatFloat(float64(d)/float64(24*time.Hour)),
+			formatFloat(float64(d)/float64(DAY)),
 		)
 
-	case d >= time.Hour:
+	case d >= HOUR:
 		return fmt.Sprintf(
 			"%.2g"+separator+"h",
-			formatFloat(float64(d)/float64(time.Hour)),
+			formatFloat(float64(d)/float64(HOUR)),
 		)
 
-	case d >= time.Minute:
+	case d >= MINUTE:
 		return fmt.Sprintf(
 			"%.2g"+separator+"m",
-			formatFloat(float64(d)/float64(time.Minute)),
+			formatFloat(float64(d)/float64(MINUTE)),
 		)
 
-	case d >= time.Second:
+	case d >= SECOND:
 		return fmt.Sprintf(
 			"%.2g"+separator+"s",
-			formatFloat(float64(d)/float64(time.Second)),
+			formatFloat(float64(d)/float64(SECOND)),
 		)
 
-	case d >= time.Millisecond:
+	case d >= MS:
 		return fmt.Sprintf(
 			"%g"+separator+"ms",
-			formatFloat(float64(d)/float64(time.Millisecond)),
+			formatFloat(float64(d)/float64(MS)),
 		)
 
-	case d >= time.Microsecond:
+	case d >= US:
 		return fmt.Sprintf(
 			"%g"+separator+"μs",
-			formatFloat(float64(d)/float64(time.Microsecond)),
+			formatFloat(float64(d)/float64(US)),
 		)
 
 	default: