+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 ## Changelog
 
+### [13.16.0](https://kaos.sh/ek/13.16.0)
+
+- `[i18n]` Added method for bundle validation `ValidateBundle`
+- `[timeutil]` Added helper `IsWeekend`
+- `[timeutil]` Added helper `EndOfHour`
+- `[timeutil]` Added helper `EndOfDay`
+- `[timeutil]` Added helper `EndOfWeek`
+- `[timeutil]` Added helper `EndOfMonth`
+- `[timeutil]` Added helper `EndOfYear`
+- `[timeutil]` Added helper `Until`
+- `[i18n]` Code refactoring
+- `[timeutil]` Code refactoring
+
 ### [13.15.12](https://kaos.sh/ek/13.15.12)
 
 - `[sliceutil]` Added method `Diff`